moved methods for adding reinforcement into geometry module
