+++ structuralcodes/sections/__init__.py
@@ -1,7 +1,6 @@
 """Main entry point for sections."""
 
 from ._generic import GenericSection, GenericSectionCalculator
-from ._reinforcement import add_reinforcement, add_reinforcement_line
 from .section_integrators import (
     FiberIntegrator,
     MarinIntegrator,
@@ -13,8 +12,6 @@ from .section_integrators import (
 __all__ = [
     'GenericSection',
     'GenericSectionCalculator',
-    'add_reinforcement',
-    'add_reinforcement_line',
     'SectionIntegrator',
     'FiberIntegrator',
     'MarinIntegrator',
